+++ src/math/mod.rs
@@ -3,14 +3,12 @@
 
 mod matrix_4;
 mod quaternion;
-mod transform;
 mod vector_2;
 mod vector_3;
 mod vector_4;
 
 pub use matrix_4::Matrix4;
 pub use quaternion::Quaternion;
-pub use transform::Transform;
 pub use vector_2::Vector2;
 pub use vector_3::Vector3;
 pub use vector_4::Vector4;